+++ sdk/servicebus/service-bus/src/serviceBusClient.ts
@@ -8,7 +8,7 @@ import {
   createConnectionContextForTokenCredential
 } from "./constructorHelpers";
 import { ConnectionContext } from "./connectionContext";
-import { CreateReceiverOptions, AcceptSessionOptions, ReceiveMode } from "./models";
+import { ServiceBusReceiverOptions, ServiceBusSessionReceiverOptions, ReceiveMode } from "./models";
 import { ServiceBusReceiver, ServiceBusReceiverImpl } from "./receivers/receiver";
 import {
   ServiceBusSessionReceiver,
@@ -124,7 +124,7 @@ export class ServiceBusClient {
    */
   createReceiver(
     queueName: string,
-    options?: CreateReceiverOptions<"peekLock">
+    options?: ServiceBusReceiverOptions<"peekLock">
   ): ServiceBusReceiver<ServiceBusReceivedMessageWithLock>;
   /**
    * Creates a receiver for an Azure Service Bus queue in receiveAndDelete mode. No connection is made
@@ -143,7 +143,7 @@ export class ServiceBusClient {
    */
   createReceiver(
     queueName: string,
-    options: CreateReceiverOptions<"receiveAndDelete">
+    options: ServiceBusReceiverOptions<"receiveAndDelete">
   ): ServiceBusReceiver<ServiceBusReceivedMessage>;
   /**
    * Creates a receiver for an Azure Service Bus subscription in peekLock mode. No connection is made
@@ -175,7 +175,7 @@ export class ServiceBusClient {
   createReceiver(
     topicName: string,
     subscriptionName: string,
-    options?: CreateReceiverOptions<"peekLock">
+    options?: ServiceBusReceiverOptions<"peekLock">
   ): ServiceBusReceiver<ServiceBusReceivedMessageWithLock>;
   /**
    * Creates a receiver for an Azure Service Bus subscription in receiveAndDelete mode. No connection is made
@@ -197,15 +197,15 @@ export class ServiceBusClient {
   createReceiver(
     topicName: string,
     subscriptionName: string,
-    options: CreateReceiverOptions<"receiveAndDelete">
+    options: ServiceBusReceiverOptions<"receiveAndDelete">
   ): ServiceBusReceiver<ServiceBusReceivedMessage>;
   createReceiver(
     queueOrTopicName1: string,
     optionsOrSubscriptionName2?:
-      | CreateReceiverOptions<"receiveAndDelete">
-      | CreateReceiverOptions<"peekLock">
+      | ServiceBusReceiverOptions<"receiveAndDelete">
+      | ServiceBusReceiverOptions<"peekLock">
       | string,
-    options3?: CreateReceiverOptions<"receiveAndDelete"> | CreateReceiverOptions<"peekLock">
+    options3?: ServiceBusReceiverOptions<"receiveAndDelete"> | ServiceBusReceiverOptions<"peekLock">
   ):
     | ServiceBusReceiver<ServiceBusReceivedMessage>
     | ServiceBusReceiver<ServiceBusReceivedMessageWithLock> {
@@ -284,7 +284,7 @@ export class ServiceBusClient {
   acceptSession(
     queueName: string,
     sessionId: string,
-    options?: AcceptSessionOptions<"peekLock">
+    options?: ServiceBusSessionReceiverOptions<"peekLock">
   ): Promise<ServiceBusSessionReceiver<ServiceBusReceivedMessageWithLock>>;
   /**
    * Creates a receiver for a session enabled Azure Service Bus queue in receiveAndDelete mode.
@@ -300,7 +300,7 @@ export class ServiceBusClient {
   acceptSession(
     queueName: string,
     sessionId: string,
-    options: AcceptSessionOptions<"receiveAndDelete">
+    options: ServiceBusSessionReceiverOptions<"receiveAndDelete">
   ): Promise<ServiceBusSessionReceiver<ServiceBusReceivedMessage>>;
   /**
    * Creates a receiver for a session enabled Azure Service Bus subscription in peekLock mode.
@@ -329,7 +329,7 @@ export class ServiceBusClient {
     topicName: string,
     subscriptionName: string,
     sessionId: string,
-    options?: AcceptSessionOptions<"peekLock">
+    options?: ServiceBusSessionReceiverOptions<"peekLock">
   ): Promise<ServiceBusSessionReceiver<ServiceBusReceivedMessageWithLock>>;
   /**
    * Creates a receiver for a session enabled Azure Service Bus subscription in receiveAndDelete mode.
@@ -347,19 +347,21 @@ export class ServiceBusClient {
     topicName: string,
     subscriptionName: string,
     sessionId: string,
-    options: AcceptSessionOptions<"receiveAndDelete">
+    options: ServiceBusSessionReceiverOptions<"receiveAndDelete">
   ): Promise<ServiceBusSessionReceiver<ServiceBusReceivedMessage>>;
   async acceptSession(
     queueOrTopicName1: string,
     optionsOrSubscriptionNameOrSessionId2?:
-      | AcceptSessionOptions<"peekLock">
-      | AcceptSessionOptions<"receiveAndDelete">
+      | ServiceBusSessionReceiverOptions<"peekLock">
+      | ServiceBusSessionReceiverOptions<"receiveAndDelete">
       | string,
     optionsOrSessionId3?:
-      | AcceptSessionOptions<"peekLock">
-      | AcceptSessionOptions<"receiveAndDelete">
+      | ServiceBusSessionReceiverOptions<"peekLock">
+      | ServiceBusSessionReceiverOptions<"receiveAndDelete">
       | string,
-    options4?: AcceptSessionOptions<"peekLock"> | AcceptSessionOptions<"receiveAndDelete">
+    options4?:
+      | ServiceBusSessionReceiverOptions<"peekLock">
+      | ServiceBusSessionReceiverOptions<"receiveAndDelete">
   ): Promise<
     | ServiceBusSessionReceiver<ServiceBusReceivedMessage>
     | ServiceBusSessionReceiver<ServiceBusReceivedMessageWithLock>
@@ -370,8 +372,8 @@ export class ServiceBusClient {
     let entityPath: string;
     let receiveMode: "peekLock" | "receiveAndDelete";
     let options:
-      | AcceptSessionOptions<"peekLock">
-      | AcceptSessionOptions<"receiveAndDelete">
+      | ServiceBusSessionReceiverOptions<"peekLock">
+      | ServiceBusSessionReceiverOptions<"receiveAndDelete">
       | undefined;
 
     if (
@@ -453,7 +455,7 @@ export class ServiceBusClient {
    */
   acceptNextSession(
     queueName: string,
-    options?: AcceptSessionOptions<"peekLock">
+    options?: ServiceBusSessionReceiverOptions<"peekLock">
   ): Promise<ServiceBusSessionReceiver<ServiceBusReceivedMessageWithLock>>;
   /**
    * Creates a receiver for the next available session in a session-enabled Azure Service Bus queue in receiveAndDelete mode.
@@ -467,7 +469,7 @@ export class ServiceBusClient {
    */
   acceptNextSession(
     queueName: string,
-    options: AcceptSessionOptions<"receiveAndDelete">
+    options: ServiceBusSessionReceiverOptions<"receiveAndDelete">
   ): Promise<ServiceBusSessionReceiver<ServiceBusReceivedMessage>>;
   /**
    * Creates a receiver for the next available session in a session-enabled Azure Service Bus subscription in peekLock mode.
@@ -494,7 +496,7 @@ export class ServiceBusClient {
   acceptNextSession(
     topicName: string,
     subscriptionName: string,
-    options?: AcceptSessionOptions<"peekLock">
+    options?: ServiceBusSessionReceiverOptions<"peekLock">
   ): Promise<ServiceBusSessionReceiver<ServiceBusReceivedMessageWithLock>>;
   /**
    * Creates a receiver for the next available session in a session-enabled Azure Service Bus subscription in receiveAndDelete mode.
@@ -510,15 +512,17 @@ export class ServiceBusClient {
   acceptNextSession(
     topicName: string,
     subscriptionName: string,
-    options: AcceptSessionOptions<"receiveAndDelete">
+    options: ServiceBusSessionReceiverOptions<"receiveAndDelete">
   ): Promise<ServiceBusSessionReceiver<ServiceBusReceivedMessage>>;
   async acceptNextSession(
     queueOrTopicName1: string,
     optionsOrSubscriptionName2?:
-      | AcceptSessionOptions<"peekLock">
-      | AcceptSessionOptions<"receiveAndDelete">
+      | ServiceBusSessionReceiverOptions<"peekLock">
+      | ServiceBusSessionReceiverOptions<"receiveAndDelete">
       | string,
-    options3?: AcceptSessionOptions<"peekLock"> | AcceptSessionOptions<"receiveAndDelete">
+    options3?:
+      | ServiceBusSessionReceiverOptions<"peekLock">
+      | ServiceBusSessionReceiverOptions<"receiveAndDelete">
   ): Promise<
     | ServiceBusSessionReceiver<ServiceBusReceivedMessage>
     | ServiceBusSessionReceiver<ServiceBusReceivedMessageWithLock>